3D Objects & Mobile App
As of right now it converts an entire scene into JSON form using three.js's toJSON() for 3D Objects. It then adds them to the local DOM tree and pushes that tree to the store for other clients to access. On the receiving end each client watches for an update at /store/:namespace/:object. When it's updated, each client recreates the same DOM tree the host has and uses that tree to recreate the scene with three.js's ObjectLoader.